Improving comments and removing unnecessary files
