+++ public/files/bn_methylation.bed
@@ -1,42 +1,3 @@
-<<<<<<< HEAD
-=======
-N0	0	1000000	86
-N0	1000001	2000000	85
-N0	2000001	3000000	85
-N0	3000001	4000000	84
-N0	4000001	5000000	83
-N0	5000001	6000000	85
-N0	6000001	7000000	85
-N0	7000001	8000000	83
-N0	8000001	9000000	79
-N0	9000001	10000000	82
-N0	10000001	11000000	85
-N0	11000001	12000000	84
-N0	12000001	13000000	87
-N0	13000001	14000000	88
-N0	14000001	15000000	88
-N0	15000001	16000000	89
-N0	16000001	17000000	89
-N0	17000001	18000000	86
-N0	18000001	19000000	3
-N0	19000001	20000000	6
-N0	20000001	21000000	80
-N0	21000001	22000000	27
-N0	22000001	23000000	31
-N0	23000001	24000000	12
-N0	24000001	25000000	10
-N0	25000001	26000000	60
-N0	26000001	27000000	86
-N0	27000001	28000000	86
-N0	28000001	29000000	90
-N0	29000001	30000000	86
-N0	30000001	31000000	81
-N0	31000001	32000000	87
-N0	32000001	33000000	85
-N0	33000001	34000000	88
-N0	34000001	35000000	90
-N0	35000001	36000000	87
->>>>>>> taylan
 N1	0	1000000	20
 N1	1000001	2000000	19
 N1	2000001	3000000	19
