+++ src/App.jsx
@@ -1,6 +1,5 @@
 
 import React from 'react';
-import ReactDOM from 'react-dom/client';
 import { Container } from './components';
 import { Provider } from 'react-redux';
 import { store } from './redux/store';
@@ -10,16 +9,12 @@ import { HTML5Backend } from 'react-dnd-html5-backend'
 import { Dashboard, Demo, Documentation, Home, NotFound, AgricultureDemo } from './pages';
 import CssBaseline from '@mui/material/CssBaseline';
 import { ThemeProvider, createTheme } from '@mui/material/styles';
-import { selectTheme } from 'themeSlice';
-import { useSelector } from 'react-redux';
 import { useState } from 'react';
 
 
 
 
 export default function App() {
-
-    // const themeSelector = useSelector(selectTheme)
 
     const [ isDark, setIsDark ] = useState(false)
 
@@ -36,9 +31,9 @@ export default function App() {
         },
       });
 
+      
     return (
 <>
-
         {/* // <React.StrictMode> */}
 <ThemeProvider theme={isDark ? darkTheme : lightTheme}>
 <CssBaseline />
