+++ src/components/tracks/TrackLinks.jsx
@@ -8,19 +8,17 @@ const TrackLinks = ({index, id,...props}) =>{
     const trackSelector =  useSelector(selectBasicTracks)
     const indexSelector = useSelector(selectDraggables)
 
+    // These should have all the information from the tracks, including zoom level + offset
+    let aboveData = trackSelector[indexSelector[index - 1]]
+    let belowData = trackSelector[indexSelector[index + 1]]
 
+    // Just finding these values for the returned example div below
     let aboveArray = trackSelector[indexSelector[index - 1]] ? trackSelector[indexSelector[index - 1]].array : []
     let belowArray =  trackSelector[indexSelector[index + 1]] ? trackSelector[indexSelector[index + 1]].array : []
-
-    // debugger
-
     let aboveLength = aboveArray.length 
     let aboveCap = aboveLength > 0 ?  Math.max(...aboveArray.map(d=> d.end)) : 0
-
     let belowLength = belowArray.length
     let belowCap = belowLength > 0 ? Math.max(...belowArray.map(d=> d.end)) : 0
-
-
 
     return(
     <>
