+++ src/features/draggable/DragPreview.js
@@ -1,19 +1,15 @@
 import { memo } from 'react'
-import Miniview from 'features/miniview/Miniview.js'
 import { useSelector } from 'react-redux'
-import { selectMiniviews } from 'features/miniview/miniviewSlice'
-import { teal } from '@mui/material/colors';
-import DragHandleIcon from '@mui/icons-material/DragHandle';
-import { IconButton, Button } from "@mui/material"
 import { nanoid } from '@reduxjs/toolkit';
 import BasicTrack from 'components/tracks/BasicTrack';
 import { selectBasicTracks } from 'components/tracks/basicTrackSlice';
 
 
-
+/* Essentially just returns a smaller version of a basic track when dragging
+TODO make work with the new histogram
+*/
 export const DragPreview = memo(function DragPreview({ item, groupID, width, height, className }) {
 
-    
     const basicTrackSelector = useSelector(selectBasicTracks)
 
         return (
