--- src/features/draggable/Draggable.css
@@ -1,22 +1,0 @@
-/* 
-  .handle {
-    width: 2%;
-    float: left;
-    height: 100%;
-    margin: 0%;
-    padding: 0%;
-    cursor: grab;
-  }
-  .handle_image {
-      position: relative;
-      top: 0%;
-      left: 0%;
-      margin: 0%;
-      padding: 0%; 
-      font-size: xx-large;
-      text-align: left;
-      width: 100%;
-      height: 100%;
-    
-   } */
-
