--- src/features/miniview/Miniview.css
@@ -1,5 +1,0 @@
-
-/* Placeholder for dimensions and locations */
-.miniview {
-  cursor: crosshair;
-}
