+++ src/features/parsers/gffParser.js
@@ -1,41 +1,41 @@
 
-import { text } from "d3-request"
+import { text } from "d3-fetch"
 
 
-/**
- * These functions are pulled from the previous gene density maps, need to be updated for the GUTB
- */
- 
-let info,  temporary, savedDensityData
+async function parseGFF(demoFile){
+       
+    let x = text(demoFile).then(data =>{
+        let temporary = data.split(/\n/)
+        let dataset = {}
+        let trackType = 'default'
 
-// Pulls collinearity info and makes links
-function pullGeneInfo(collinearityFile, nomenclature) {
+        // BED file processor for methylation data
+        if (demoFile.indexOf('.bed') > -1) {
 
-    return text(collinearityFile).then(function (data) {
-        let rawCollinearity = process(data);
-        let selectedCollinearity = []
-        nomenclature.forEach(n => {
-            let temporaryCollinearity = rawCollinearity.alignmentList.filter((d) => d.source.indexOf(n) > -1 && d.target.indexOf(n) > -1)
-            selectedCollinearity.push(...temporaryCollinearity)
-        })
-        let genePairs = selectedCollinearity.reduce((c, e) => { return [...c, ...e.links] }, [])
-        let trueMatch = genePairs.filter((x) => +x.e_value == 0)
+            trackType = 'histogram'
 
-        return trueMatch
-    })
-}
+            temporary.forEach(d => {
+                let info = d.split('\t')
+                if (info.length > 1) {
+                    let key = info[0] + "-" + info[1] + "-" + info[2]
+                    var stats = {
+                        chromosome: info[0],
+                        start: info[1],
+                        end: info[2],
+                        key,
+                        ortholog: false,
+                        siblings: [],
+                        value: info[3]
+                    }
+                    dataset[key] = stats
+                }
+            })
+        }
 
-// Turns gff file into model
-export function pullInfo(genomeFile) {
-    return text(genomeFile).then(function (data) {
-            temporary = data.split(/\n/)
-            // dataset = []
-            savedDensityData = []
-
-            let dataset = {}
-            temporary.forEach(d=>{
-                info = d.split('\t')
-                // console.log(info)
+        // gff file parser
+        else {
+            temporary.forEach(d => {
+                let info = d.split('\t')
                 if (info.length > 1) {
                     let key = info[1].toLowerCase()
                     var stats = {
@@ -45,90 +45,59 @@ export function pullInfo(genomeFile) {
                         key: key,
                         ortholog: false,
                         siblings: [],
+                        value: 0
                     }
-                    // console.log(stats)
-                    // console.log(key)
-                   dataset[key] = stats
-            }
-        })
-            return dataset
-        })
-    }
-
-
-function process(collinearityData) {
-    // The first 11 lines contain information regarding the MCSCANX Parameters
-    // and can be processed seperately 
-    var FileLines = collinearityData.split('\n'),
-        information = parseInformation(FileLines.slice(0, 11)),
-        alignmentList = [],
-        alignmentBuffer = {};
-    // remove the first 11 lines and then process the file line by line
-    FileLines.slice(11).forEach(function (line, index) {
-        if (line.indexOf('Alignment') > -1) {
-            // store the previous alignment in list , 
-            // and skip for the first iteration since buffer is empty
-            if (alignmentBuffer.links) {
-                alignmentList.push(alignmentBuffer);
-            }
-            alignmentBuffer = parseAlignmentDetails(line);
-            alignmentBuffer.links = [];
-        } else if (line.trim().length > 1) {
-            // condition to skip empty lines
-            alignmentBuffer.links.push(parseLink(line));
+                    dataset[key] = stats
+                }
+            })
         }
+
+        // Building up the different chromosomes
+        let chromosomeNameList = []
+        let chromosomalData = []
+        let ignore = "Scaffold"
+
+        for (let item in dataset) {
+            if (!chromosomeNameList.some((x) => x.chromosome == dataset[item].chromosome) && !dataset[item].chromosome.includes(ignore)) {
+                var check = item
+                var temp = {
+                    chromosome: dataset[item].chromosome,
+                    designation: check.slice(0, check.indexOf('g'))
+                }
+                // Building a list of the chromosome names, used for later finding information on that dataset
+                chromosomeNameList.push(temp)
+            }
+        }
+
+        // Changing the default lexicographical order, since chromosome11 should come after chromosome2 
+        // additional logic so that all chromosomes from the same line should be grouped   
+        chromosomeNameList.sort((a, b) => {
+            if (a.chromosome[0].localeCompare(b.chromosome[0]) == 0) {
+                return a.chromosome.length - b.chromosome.length
+            }
+            else {
+                return a.chromosome[0].localeCompare(b.chromosome[0])
+            }
+        })
+
+        chromosomeNameList.forEach((chr) => {
+          var subset = Object.entries(dataset).filter(d => {
+            return d[1].chromosome == chr.chromosome
+          }).map(x => x[1])
+          
+          var temp = {
+            key: chr,
+            data: subset,
+            trackType,
+          }
+          chromosomalData.push(temp)
+        })
+        
+        return {chromosomalData, dataset}
     })
-    // push the last alignment still in the buffer
-    alignmentList.push(alignmentBuffer);
-    // get the unique list of IDs of all chromosomes or scaffolds that have alignments mapped to them
-    let uniqueIDList = [];
-    alignmentList.map((d) => { uniqueIDList.push(d.source, d.target) });
-    return { "information": information, "alignmentList": alignmentList, 'uniqueIDList': uniqueIDList.filter(onlyUnique) };
-};
 
+    return x
 
-function parseInformation(informationLines) {
-    return {
-        'parameters': [
-            ['match score', informationLines[1].split(':')[1].trim()],
-            ['match size', informationLines[2].split(':')[1].trim()],
-            ['gap penality', informationLines[3].split(':')[1].trim()],
-            ['overlap wndow', informationLines[4].split(':')[1].trim()],
-            ['e value', informationLines[5].split(':')[1].trim()],
-            ['maximum gaps', informationLines[6].split(':')[1].trim()]
-        ],
-        'stats': {
-            'no_of_collinear_genes': informationLines[8].split(',')[0].split(":")[1].trim(),
-            'percentage': Number(informationLines[8].split(',')[1].split(":")[1].trim()),
-            'no_of_all_genes': informationLines[8].split(',')[1].split(":")[1].trim()
-        }
-    };
 }
 
-function parseAlignmentDetails(alignmentDetails) {
-    let alignmentDetailsList = alignmentDetails.split(' ');
-    return {
-        'score': Number(alignmentDetailsList[3].split('=')[1].trim()),
-        'e_value': Number(alignmentDetailsList[4].split('=')[1].trim()),
-        'count': Number(alignmentDetailsList[5].split('=')[1].trim()),
-        'type': alignmentDetailsList[7].trim() == 'plus' ? 'regular' : 'flipped',
-        'source': alignmentDetailsList[6].split('&')[0].trim(),
-        'target': alignmentDetailsList[6].split('&')[1].trim(),
-        'sourceKey': Number(alignmentDetailsList[6].split('&')[0].trim().slice(2)),
-        'targetKey': Number(alignmentDetailsList[6].split('&')[1].trim().slice(2)),
-        'alignmentID': Number(alignmentDetailsList[2].split(':')[0].trim())
-    };
-}
-
-function parseLink(link) {
-    let linkInfo = link.split('\t');
-    return {
-        'source': linkInfo[1],
-        'target': linkInfo[2],
-        'e_value': linkInfo[3]
-    };
-}
-
-function onlyUnique(value, index, self) {
-    return self.indexOf(value) === index;
-}
+export default parseGFF
