+++ src/redux/store.js
@@ -13,6 +13,5 @@ export const store = configureStore({
     alternateDraggable: alternateDraggableReducer,
     draggable: draggableReducer,
     basictrack : basicTrackReducer,
-
   },
 });
